Move My Spaces to own page
