+++ src/PersonalWebSpace_frontend/pages/Intro.svelte
@@ -5,40 +5,6 @@
   import Button from "../components/Button.svelte";
   import Topnav from "../components/Topnav.svelte";
   import Footer from "../components/Footer.svelte";
-  import UserSpaces from "../components/UserSpaces.svelte";
-
-  let hasLoadedSpaces = false;
-  let loadedUserSpaces = [];
-
-  const createNewUserSpace = async (element) => {
-    element.setAttribute("disabled", true);
-    document.getElementById("createSubtext").innerText = "Creating your Personal Web Space, just a moment...";
-    
-    const resp = await fetch("defaultRoom.html"); // Fetches default space each user gets initially
-    const defaultSpaceHtml = await resp.text();
-    const space = await $store.backendActor.createSpace(defaultSpaceHtml);
-    document.getElementById("createSubtext").innerText = "Ohh yeah, you just got yourself a new Personal Web Space!";
-    
-    // Reload user's spaces
-    loadUserSpaces();
-
-    element.removeAttribute("disabled");
-  };
-
-  const loadUserSpaces = async () => {
-    const userSpaces = await $store.backendActor.getCallerSpaces();
-    const numberOfSpacesUserOwns = userSpaces.length;
-    if (numberOfSpacesUserOwns < 1) {
-      document.getElementById("spacesSubtext").innerText = "You don't own any spaces yet. Get your Personal Web Space now by clicking on the Create tab!";
-    } else {
-      document.getElementById("spacesSubtext").innerText = numberOfSpacesUserOwns === 1 
-        ? `Big success, you own ${numberOfSpacesUserOwns} space! Let's take a look:`
-        : `Big success, you own ${numberOfSpacesUserOwns} spaces! Let's take a look:`;
-
-      loadedUserSpaces = userSpaces;
-      hasLoadedSpaces = true;
-    }
-  };
 </script>
 
 <Topnav />
@@ -55,33 +21,27 @@
   {:else}
     <h3 class="font-bold">You're Logged In</h3>
     <div>Principal: {$store.principal}</div>
-    <div>AccountId: {$store.accountId}</div>
     <Button on:click={() => store.disconnect()}>disconnect</Button>
   {/if}
 </section>
 
 <section id="create" class="py-7 space-y-6 items-center text-center">
   <h3 class="font-bold">Create a new Personal Web Space</h3>
+  <button type='button' id='createButton' on:click={() => push("#/create")} class="active-app-button bg-slate-500 text-white font-bold py-2 px-4 rounded">Create Space</button>
   {#if !$store.isAuthed}
-    <button type='button' id='createButton' disabled class="bg-slate-500 text-white font-bold py-2 px-4 rounded opacity-50 cursor-not-allowed">Create Space</button>
     <p id='createSubtext'>Log in to generate a 3D room (Your Space, Your Realm, Your Virtual Home) which you can edit afterwards. Fun fact: The Space is an NFT itself and will be sent to your wallet. This way you know it's truly yours!</p>
   {:else}
-    <button type='button' id='createButton' on:click={() => push("#/create")} class="active-app-button bg-slate-500 text-white font-bold py-2 px-4 rounded">Create Space</button>
-    <p id='createSubtext'>Click and we'll generate a 3D room for you (Your Space, Your Realm, Your Virtual Home) which you can edit afterwards. Fun fact: The Space is an NFT itself and will be sent to your wallet. This way you know it's truly yours!</p>
+    <p id='createSubtext'>Generate a 3D room for yourself (Your Space, Your Realm, Your Virtual Home) which you can edit afterwards. Fun fact: The Space is an NFT itself and will be sent to your wallet. This way you know it's truly yours!</p>
   {/if}
 </section>
 
 <section id="spaces" class="py-7 space-y-6 items-center text-center bg-slate-100">
-  <h3 class="font-bold">My Personal Web Spaces</h3>
+  <h3 class="font-bold">See My Personal Web Spaces</h3>
+  <button type='button' id='viewMySpacesButton' on:click={() => push("#/myspaces")} class="active-app-button bg-slate-500 text-white font-bold py-2 px-4 rounded">View My Spaces</button>
   {#if !$store.isAuthed}
-    <p id='spacesSubtext'>Log in to see which Spaces you own.</p>
+    <p id='spacesSubtext'>Log in to see which Spaces you own in the Open Internet Metaverse.</p>
   {:else}
-    <p id='spacesSubtext'>Let's see which Spaces you own...</p>
-    {#if !hasLoadedSpaces}
-      <p hidden>{loadUserSpaces()}</p>
-    {:else}
-      <UserSpaces spaces={loadedUserSpaces} />
-    {/if}
+    <p id='spacesSubtext'>See which Spaces you own in the Open Internet Metaverse.</p>
   {/if}
 </section>
